refactor(month picker): and fix KB navigation performance in IE #3126
