+++ projects/igniteui-angular/src/lib/calendar/calendar.component.html
@@ -38,9 +38,9 @@
                    [animationAction]="monthAction"
                    [locale]="locale"
                    [value]="value"
+                   [viewDate]="viewDate"
                    [weekStart]="weekStart"
                    [formatOptions]="formatOptions"
-                   [viewDate]="viewDate"
                    [selection]="selection"
                    [disabledDates]="disabledDates"
                    [specialDates]="specialDates"
