+++ projects/igniteui-angular/src/lib/calendar/days-view/day-item.component.html
@@ -1,3 +1,1 @@
-<span>
-    <ng-content></ng-content>
-</span>
+<ng-content></ng-content>
