+++ projects/igniteui-angular/src/lib/calendar/index.ts
@@ -3,5 +3,6 @@ export * from './calendar.component';
 export * from './days-view/days-view.component';
 export * from './months-view/months-view.component';
 export * from './years-view/years-view.component';
+export * from './month-picker/month-picker.component';
 export * from './calendar.directives';
 export * from './calendar.module';
