+++ projects/igniteui-angular/src/lib/calendar/years-view/years-view.component.html
@@ -1,6 +1,6 @@
 <div class="igx-calendar__body">
     <div class="igx-calendar__body-column" (wheel)="scroll($event)" (pan)="pan($event)">
-        <span [igxCalendarYear]="year" [date]="date" (onYearSelection)="selectYear($event)" *ngFor="let year of decade">
+        <span [igxCalendarYear]="year" [date]="date" (onYearSelection)="selectYear($event)" *ngFor="let year of decade; trackBy: yearTracker">
             {{ formattedYear(year) }}
         </span>
     </div>
