+++ projects/igniteui-angular/src/public_api.ts
@@ -60,7 +60,6 @@ export * from './lib/checkbox/checkbox.component';
 export * from './lib/chips/index';
 export * from './lib/combo/index';
 export * from './lib/date-picker/date-picker.component';
-export * from './lib/month-picker/month-picker.component';
 export * from './lib/dialog/dialog.component';
 export * from './lib/drop-down/index';
 export * from './lib/grids/grid/index';
